+++ src/InputHandler.h
@@ -88,8 +88,8 @@ class InputHandler : public XMLHandler
 		string getOutputPlugin(string);
 		string getFileName(string type);
 		string getResultFileName();
-		int    getRegion(string direction);
-		int    getGridSize();
+		double getRegion(string direction);
+		double getGridSize();
 		unsigned long int getTimeSteps();
         unsigned long int getTimeIncrement();
 		DOMNode* getExperiment();
